mazeSolver/requirements.txt
matplotlib==3.3.2
opencv_python==4.4.0.46
numpy==1.18.5
Pillow==8.2.0
